Fix slope sign error in tabular SFH bin weight calculations.
This bug caused potentially major inaccuracies in output magnitudes
and other physical quantities for tabular SFHs, by incorrectly renormalizing
the SSP weights for each bin in the tabular SFH when the slope of the
linearly interpolated SFR was not zero. SFHs with larger differences
between adjacent SFRs and/or poorer temporal sampling, and bands dominated by
younger stars are more strongly affected.